fix: resolve issues with solutions updates on answer collection edit
