+++ apps/frontend-manage/src/components/questions/manipulation/options/CaseStudyManualItemCreation.tsx
@@ -156,13 +156,6 @@ function CaseStudyManualItemCreation({
           // reset the manually created items
           manualItemsHelpers.setValue([])
 
-          // manually reset the sample solutions defined for the created cases
-          const newCases = casesField.value?.map((caseItem) => ({
-            ...caseItem,
-            solutions: undefined,
-          }))
-          casesHelpers.setValue(newCases)
-
           // switch to the selection mode for existing answer collections
           setItemSelectionMode('existing')
         }}
